+++ plugins/orchestrator-backend/CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Dependencies
 
+* **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.21.0
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.20.0
 * **@janus-idp/cli:** upgraded to 1.15.2